coverted period length to integer

+++ src/main/kotlin/uk/gov/justice/digital/hmpps/hmppsremandandsentencingapi/controller/dto/PeriodLength.kt
@@ -1,13 +1,12 @@
 package uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.controller.dto
 
 import uk.gov.justice.digital.hmpps.hmppsremandandsentencingapi.jpa.entity.PeriodLengthEntity
-import java.math.BigDecimal
 
 data class PeriodLength(
-  val years: BigDecimal?,
-  val months: BigDecimal?,
-  val weeks: BigDecimal?,
-  val days: BigDecimal?,
+  val years: Int?,
+  val months: Int?,
+  val weeks: Int?,
+  val days: Int?,
   val periodOrder: String,
 ) {
   companion object {
